% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualEvaporation.R
\name{actualEvaporation}
\alias{actualEvaporation}
\title{Calculate Actual Evapotranspiration for Impervious Areas}
\usage{
actualEvaporation(
  usageTuple,
  potentialEvaporation,
  soilProperties,
  precipitation,
  log = TRUE
)
}
\arguments{
\item{usageTuple}{list as returned by \code{\link{getUsageTuple}}}

\item{potentialEvaporation}{potential evaporation in mm}

\item{soilProperties}{list as returned by \code{\link{getSoilProperties}}}

\item{precipitation}{precipitation in mm}

\item{log}{logical indicating whether or not to show debug messages}
}
\description{
Calculate Actual Evapotranspiration for Impervious Areas
}
