% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{BERLIN_TYPES_TO_LAND_TYPE_VEG_CLASS_IRRIGATION}
\alias{BERLIN_TYPES_TO_LAND_TYPE_VEG_CLASS_IRRIGATION}
\title{Assignment between (NUTZUNG, TYP) and (land_type, veg_class, irrigation)}
\format{
An object of class \code{data.frame} with 181 rows and 5 columns.
}
\usage{
BERLIN_TYPES_TO_LAND_TYPE_VEG_CLASS_IRRIGATION
}
\description{
The following tables are read from csv files and then merged:
}
\details{
berlin_tuples.csv
table of different occurring (land_type, veg_class, irrigation) tuples

berlin_type_tuple_groups.csv
assignments between berlin_type (input column "TYP") and tuples

berlin_usage_to_type_tuple_groups.csv
assignments between berlin_usage (input column "NUTZUNG") and
assignments between berlin_type (input column "TYP") and tuple
}
\keyword{datasets}
