% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-scenario.R
\name{calculate_delta_w_1}
\alias{calculate_delta_w_1}
\title{Deviation from Natural Water Balance (Delta-W)}
\usage{
calculate_delta_w_1(
  natural,
  urban,
  cols_to_omit = c("area"),
  return_codes = FALSE
)
}
\arguments{
\item{natural}{R-Abimo results for the natural scenario}

\item{urban}{R-Abimo results for the "urban" scenario}

\item{cols_to_omit}{column names that not contain result data or code identifiers. Defaults to "total_area"}

\item{return_codes}{a logical value determining whether the codes should be returned along the delta-w values}
}
\value{
a dataframe containing the delta-w values (and optionally the areas' codes)
}
\description{
Calculate the deviation from the natural water balance (delta-W) given
R-Abimo results as returned by \code{\link{run_rabimo}}.
}
