% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_with_data.R
\name{call_with_data}
\alias{call_with_data}
\title{Call a Function with Argument Combinations from a Data Frame}
\usage{
call_with_data(FUN, data, ..., threshold = 0.5)
}
\arguments{
\item{FUN}{function to be called}

\item{data}{data frame with one column per argument of \code{FUN}}

\item{\dots}{further arguments passed to \code{\link{mapply}}}

\item{threshold}{if the ratio of unique value combinations in the relevant
columns in data to all value combinations in these columns is below this
threshold value then FUN will be called only with the unique value
combinations. This should increase performance.}
}
\value{
vector of length \code{nrow(data)} with the result values returned by
\code{FUN}
}
\description{
Call a Function with Argument Combinations from a Data Frame
}
\examples{
combis <- expand.grid(x = 1:2, y = c(10, 20, 30))
combis

call_with_data(`+`, combis)
}
