% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{crop_box}
\alias{crop_box}
\title{Crop a box out of a shape}
\usage{
crop_box(x, scale = 0.02, xshift = 0, yshift = 0)
}
\arguments{
\item{x}{sf object}

\item{scale}{fraction of total width/height of x to be taken}

\item{xshift}{can be used to shift the box to the left/right}

\item{yshift}{can be used to shift the box to the bottom/top}
}
\description{
Crop a box out of a shape
}
