% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{define_controls}
\alias{define_controls}
\title{Define List of "Controls"}
\usage{
define_controls(check = TRUE, simulate_abimo = TRUE, intermediates = FALSE)
}
\arguments{
\item{check}{logical indicating whether the check functions are executed.
Default: \code{TRUE}.}

\item{simulate_abimo}{logical indicating whether or not to simulate exactly
what Abimo does (including obvious errors!). Default: \code{TRUE}!.}

\item{intermediates}{logical indicating whether the intermediate results are
returned as attributes. Default: \code{FALSE}.}
}
\value{
list with the arguments of this function as list elements
}
\description{
Define a list of settings that control how the main function
\code{\link{run_rabimo}} should behave.
}
