% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure_stats.R
\name{get_measure_stats}
\alias{get_measure_stats}
\title{Get Mean/Max Statistics on Measures}
\usage{
get_measure_stats(blocks, reference_system = 2)
}
\arguments{
\item{blocks}{data frame similar to \code{\link{rabimo_inputs_2020}$data},
with each row representing a block area}

\item{reference_system}{indicator for the "reference system" in which the
returned values are to be given. 1: all values refer to percentages of
specific areas (green roof: roof area, unsealed: total area, to_swale:
sealed area); 2: all values refer to percentages of the total area. The
default is 2.}
}
\value{
list with elements "mean" and "max" each of which is a list with one
element per measure
}
\description{
Get Mean/Max Statistics on Measures
}
