% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_potential_evaporation.R
\name{get_potential_evaporation}
\alias{get_potential_evaporation}
\title{Provide Data on Potential Evaporation}
\usage{
get_potential_evaporation(is_waterbody, district, lookup)
}
\arguments{
\item{is_waterbody}{(vector of) logical indicating whether a block area is
of type (from the type/yield/irrigation tuple) "waterbody"}

\item{district}{(vector of) integer indicating the district number of the
plot area (from the original input column "BEZIRK")}

\item{lookup}{data frame with key columns \code{is_waterbody}, \code{district}
and value columns \code{etp}, \code{etps}. A data frame of the required
structure is returned by \code{\link{abimo_config_to_config}} in list
element \code{"potential_evaporation"}}
}
\description{
Provide Data on Potential Evaporation
}
\examples{
\dontrun{
config <- abimo_config_to_config(kwb.abimo:::read_config())
get_potential_evaporation(
  is_waterbody = TRUE,
  district = 1,
  lookup = config$potential_evaporation
)

get_potential_evaporation(
  is_waterbody = c(TRUE, TRUE, FALSE, FALSE),
  district = c(1, 99, 2, 99),
  lookup = config$potential_evaporation
)
}

}
