% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precipitation.R
\name{get_precipitation}
\alias{get_precipitation}
\title{Provide Information on Precipitation}
\usage{
get_precipitation(precipitation_year, precipitation_summer, correction_factor)
}
\arguments{
\item{precipitation_year}{precipitation per year in mm}

\item{precipitation_summer}{precipitation within summer period in mm}

\item{correction_factor}{correction factor}
}
\value{
list with elements \code{year}, \code{summer}
}
\description{
Provide Information on Precipitation
}
\examples{
get_precipitation(600, 300, 0.8)
}
