% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usage_tuple.R
\name{get_usage_tuple}
\alias{get_usage_tuple}
\title{Get Usage Tuple (Usage, Yield, Irrigation) from NUTZUNG and TYP}
\usage{
get_usage_tuple(usage, type, include_inputs = FALSE)
}
\arguments{
\item{usage}{value of column NUTZUNG in input data frame}

\item{type}{value of column TYP in input data frame}

\item{include_inputs}{logical indicating whether or not to include the
input values in the output}
}
\value{
list with elements \code{usage}, \code{yield}, \code{irrigation}
}
\description{
Get Usage Tuple (Usage, Yield, Irrigation) from NUTZUNG and TYP
}
\examples{
get_usage_tuple(10, 10)
get_usage_tuple(10, 1:3)
}
