% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultConfiguration.R
\name{getDefaultConfiguration}
\alias{getDefaultConfiguration}
\title{Get ABIMO Default Configuration}
\usage{
getDefaultConfiguration(precipitationCorrectionFactor)
}
\arguments{
\item{precipitationCorrectionFactor}{precipitation correction factor}
}
\value{
list with elements
precipitationCorrectionFactor,
bagrovValues,
runoffFactors,
potentialEvaporation
}
\description{
Get ABIMO Default Configuration
}
