% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPotentialEvaporation.R
\name{getPotentialEvaporation}
\alias{getPotentialEvaporation}
\title{Provide Data on Potential Evaporation}
\usage{
getPotentialEvaporation(
  input,
  config,
  default_etp = 660L,
  default_etps = 530L,
  default_etp_waterbody = 775L
)
}
\arguments{
\item{input}{data frame with columns \code{usage}, \code{district}}

\item{config}{list with entry \code{potentialEvaporation}}

\item{default_etp}{default value for etp}

\item{default_etps}{default value for etps}

\item{default_etp_waterbody}{default value for etp for usage = 110
(waterbody)}
}
\description{
Provide Data on Potential Evaporation
}
\examples{
getPotentialEvaporation(data.frame(usage = 10, district = 1), config = list(
 potentialEvaporation = list(district_1 = list(etp = 100, etps = 200))
))

}
