% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPotentialEvaporation2.R
\name{getPotentialEvaporation2}
\alias{getPotentialEvaporation2}
\title{Provide Data on Potential Evaporation}
\usage{
getPotentialEvaporation2(isWaterbody, district, abimo_config)
}
\arguments{
\item{isWaterbody}{(vector of) logical indicating whether a block area is
of type (from the type/yield/irrigation tuple) "waterbody"}

\item{district}{(vector of) integer indicating the district number of the
plot area (from the original input column "BEZIRK")}

\item{abimo_config}{list structure as returned by
\code{kwb.abimo:::read_config}}
}
\description{
Provide Data on Potential Evaporation
}
\examples{
\dontrun{
abimo_config <- kwb.abimo:::read_config()
getPotentialEvaporation(
  is_waterbody = TRUE,
  district = 1,
  abimo_config = abimo_config
)
}

}
