% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecipitation.R
\name{getPrecipitation}
\alias{getPrecipitation}
\title{Provide Information on Precipitation}
\usage{
getPrecipitation(precipitationYear, precipitationSummer, correctionFactor)
}
\arguments{
\item{precipitationYear}{precipitation per year in mm}

\item{precipitationSummer}{precipitation within summer period in mm}

\item{correctionFactor}{correction factor}
}
\value{
list with elements \code{perYearInteger},
\code{perYearCorrectedFloat}, \code{inSummerFloat}
}
\description{
Provide Information on Precipitation
}
\examples{
getPrecipitation(600, 300, 0.8)
}
