% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilProperties.R
\name{getSoilProperties}
\alias{getSoilProperties}
\title{Calculate Soil Properties}
\usage{
getSoilProperties(
  usage,
  yield,
  depthToWaterTable,
  fieldCapacity_30,
  fieldCapacity_150
)
}
\arguments{
\item{usage}{usage string, one of "vegetationless_D", "waterbody_G",
"horticultural_K", "agricultural_L", "forested_W"}

\item{yield}{yield class}

\item{depthToWaterTable}{depth to water table}

\item{fieldCapacity_30}{field capacity in 30 cm depth}

\item{fieldCapacity_150}{field capacity in 150 cm depth}
}
\description{
Calculate Soil Properties
}
