% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_block_2.R
\name{plot_block_2}
\alias{plot_block_2}
\title{Plot Area Fractions of one Block as Part of a Square}
\usage{
plot_block_2(
  block,
  mar = c(1, 1, 3, 1),
  density_sealed = 10L,
  density_connected = 10L,
  col_main = "white",
  col_road = "lightgrey"
)
}
\arguments{
\item{block}{data frame with one row and columns as available
in the data frame that is returned by \code{\link{prepare_input_data}}}

\item{mar}{margin vector being passed to \code{\link{par}}}

\item{density_sealed}{density of shading lines indicating sealed areas}

\item{density_connected}{density of shading lines indicating connected areas}
}
\description{
Plot Area Fractions of one Block as Part of a Square
}
