% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_berlin_data.R
\name{prepare_berlin_data}
\alias{prepare_berlin_data}
\title{Function to be called when using ABIMO with the 'old' Berlin-data structure}
\usage{
prepare_berlin_data(data_file, config_file = kwb.abimo:::default_config_file())
}
\arguments{
\item{data_file}{a path for the .dbl file containing the input data}

\item{config_file}{a path for the .XML config file}
}
\value{
input a list object containing elements \code{data},
the prepared input data and \code{config}, the prepared config object,
both in a form readable by \code{\link{run_rabimo}}
}
\description{
It calls the functions \code{\link{prepare_input_data}} and
\code{\link{abimo_config_to_config}} and modifies the config-object to the required
structure.
}
