% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_berlin_inputs.R
\name{prepare_berlin_inputs}
\alias{prepare_berlin_inputs}
\title{Function to be called when using ABIMO with the 'old' Berlin-data structure}
\usage{
prepare_berlin_inputs(
  data_file,
  config_file = kwb.abimo::default_config(),
  data = NULL,
  config = NULL,
  dbg = TRUE
)
}
\arguments{
\item{data_file}{a path for the .dbf file containing the input data}

\item{config_file}{a path for the .xml config file}

\item{data}{optional. Data frame containing the input data, as e.g. provided
in \code{\link[kwb.abimo]{abimo_input_2019}}.}

\item{config}{optional. List representing an Abimo configuration, as e.g.
provided by \code{\link[kwb.abimo]{read_config}}.}

\item{dbg}{logical indicating whether or not to show debug messages}
}
\value{
a list object containing elements \code{data},
the prepared input data and \code{config}, the prepared config object,
both in a form readable by \code{\link{run_rabimo}}
}
\description{
It calls the functions \code{\link{prepare_input_data}} and
\code{\link{abimo_config_to_config}} and modifies the config-object to the
required structure.
}
