% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareInputData.R
\name{prepareInputData}
\alias{prepareInputData}
\title{Prepare Input Data: Rename, Add Columns}
\usage{
prepareInputData(input_data)
}
\arguments{
\item{input_data}{data frame with columns REGENJA, REGENSO, NUTZUNG, TYP,
BEZIRK, FLGES, STR_FLGES, PROBAU, PROVGU, VGSTRASSE, KAN_BEB, BELAG1,
BELAG2, BELAG3, BELAG4, KAN_VGU, STR_BELAG1, STR_BELAG2, STR_BELAG3,
STR_BELAG4, KAN_STR, FLUR, FELD_30, FELD_150}
}
\value{
\code{input_data} with columns renamed and additional columns
(e.g. ratios calculated from percentages)
}
\description{
Rename columns from ABIMO 3.2 original names to ABIMO 3.3 internal names
}
