% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rabimo_inputs_2020}
\alias{rabimo_inputs_2020}
\title{"Official" R-Abimo Inputs (Data and Config) for Berlin, 2020}
\format{
A list containing two elements:
1 - a data frame with the input data in r-abimo format ...(number of vars)
2 - a list object with configuration data
\describe{
\item{CODE}{Variable 'CODE'}
...
}
}
\usage{
rabimo_inputs_2020
}
\description{
The data was ...
}
\keyword{datasets}
