% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rabimo_inputs_2025}
\alias{rabimo_inputs_2025}
\title{R-Abimo Inputs (Data and Config) for Berlin, 2025}
\format{
\subsection{\code{rabimo_inputs_2025}}{

A list containing two elements:
\describe{
\item{data}{a data frame with the input data in r-abimo format ...(number of vars)}
\item{config}{a list object with configuration data}
}
}

\subsection{\code{rabimo_inputs_2025$data}}{

\describe{
\item{CODE}{Variable 'CODE'}
...
}
}

\subsection{\code{rabimo_inputs_2025$config}}{

\describe{
\item{runoff_factors}{Runoff factors}
\item{bagrov_values}{Bagrov values for sealed surfaces}
\item{swale}{Model parameter(s) related to the 'swale' measure}
}
}
}
\source{
\url{https://gdi.berlin.de/services/wfs/ua_gruendaecher_2020?REQUEST=GetCapabilities&SERVICE=wfs}
}
\usage{
rabimo_inputs_2025
}
\description{
Data frame and configuration object that are required by the R-Abimo main
function \code{\link{run_rabimo}}.
}
\keyword{datasets}
