% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_to_seq}
\alias{range_to_seq}
\title{Sequence of Values Between the Range of Values in a Given Vector}
\usage{
range_to_seq(x, by = 1)
}
\arguments{
\item{x}{vector of values from which to take the range}

\item{by}{increment of seqence}
}
\value{
sequence of values between \code{min(x)} and \code{max(x)} with
increment \code{by}
}
\description{
Sequence of Values Between the Range of Values in a Given Vector
}
