% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_evapo_transpiration.R
\name{real_evapo_transpiration}
\alias{real_evapo_transpiration}
\title{Calculate Actual Evapotranspiration with Bagrov}
\usage{
real_evapo_transpiration(
  precipitation,
  potential_evaporation,
  bagrov_parameter,
  x_ratio = NULL,
  FUN_y_ratio = y_ratio_3,
  ...
)
}
\arguments{
\item{precipitation}{precipitation in mm}

\item{potential_evaporation}{potential evaporation in mm}

\item{bagrov_parameter}{Bagrov parameter (n-value)}

\item{x_ratio}{optional. Instead of \code{precipitation} and
\code{potential_evaporation} the quotient of both may be passed to this
function. The idea is to calculate the quotient out of the function and to
reuse the quotient instead of recalculating it.}

\item{FUN_y_ratio}{function to be called to calculate the y_ratio(s) from the
given x_ratio(s). Default: \code{kwb.rabimo:::y_ratio_3}}

\item{\dots}{further arguments passed to \code{FUN_y_ratio}}
}
\value{
estimated actual evapotranspiration in mm
}
\description{
Calculate Actual Evapotranspiration with Bagrov
}
