% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realEvapoTranspiration.R
\name{realEvapoTranspiration}
\alias{realEvapoTranspiration}
\title{Calculate Actual Evapotranspiration with Bagrov}
\usage{
realEvapoTranspiration(precipitation, potentialEvaporation, efficiency)
}
\arguments{
\item{precipitation}{precipitation in mm}

\item{potentialEvaporation}{potential evaporation in mm}

\item{efficiency}{so-called Bagrov value}
}
\value{
estimated actual evapotranspiration in mm
}
\description{
Calculate Actual Evapotranspiration with Bagrov
}
