% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_api.R
\name{test_api}
\alias{test_api}
\title{Test the API to main functions using the plumber package and swagger}
\usage{
test_api()
}
\description{
Test the API to main functions using the plumber package and swagger
}
