% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plumber_api.R
\name{test_plumber_api}
\alias{test_plumber_api}
\title{Test the API to main functions using the plumber package and swagger}
\usage{
test_plumber_api()
}
\description{
Test the API to main functions using the plumber package and swagger
}
