% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_of_fractions.R
\name{triangle_of_fractions}
\alias{triangle_of_fractions}
\title{Plot Triangle of Three Fractions}
\usage{
triangle_of_fractions(
  fractions,
  fractions_2 = NULL,
  cols = c("blue", "red", "darkgreen")
)
}
\arguments{
\item{fractions}{numeric vector with three values having a sum of one. The
names of the vector elements are used as labels}

\item{fractions_2}{optional. Similar to \code{fractions}. If given, these
fractions are shown as dashed lines in the plot and the "deltas" between
\code{fractions} and \code{fractions_2} are shown as horizontally
stacked bars below the triangle.}

\item{cols}{vector of length three giving the colour names}
}
\description{
Plot Triangle of Three Fractions
}
\examples{
# blue, red, green | left, right, bottom
components <- c(runoff = 200, infiltration = 50, evaporation = 100)
fractions <- components / sum(components)
triangle_of_fractions(fractions)
triangle_of_fractions(fractions, fractions_2 = c(0.1, 0.3, 0.6))
}
