% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_evapo_transpiration.R
\name{y_ratio_3}
\alias{y_ratio_3}
\title{Lookup y_ratio for given x_ratio on a BAGROV curve}
\usage{
y_ratio_3(
  bagrov_parameter,
  x_ratio,
  min_size_for_parallel = 10L,
  use_abimo_algorithm = FALSE
)
}
\arguments{
\item{bagrov_parameter}{(vector of) BAGROV parameter(s)}

\item{x_ratio}{(vector of) x-ratio(s) (between precipitation and potential
evaporation) for which to look up the corresponding y-ratio(s) (between
actual evaporation and potential evaporation) on the BAGROV curve(s)}

\item{min_size_for_parallel}{minimum number of BAGROV curves to start
parallel processing}

\item{use_abimo_algorithm}{whether or not to use the original algorithm that
is implemented in the C++ code (converted to R:
\code{kwb.rabimo:::yratio_cpp}). Default: \code{FALSE}}
}
\description{
Lookup y_ratio for given x_ratio on a BAGROV curve
}
