% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_height_to_volume_flow.R
\name{yearly_height_to_volume_flow}
\alias{yearly_height_to_volume_flow}
\title{Convert Yearly Height (mm) to Volume Flow (unit?)}
\usage{
yearly_height_to_volume_flow(height, area)
}
\arguments{
\item{height}{height in mm}

\item{area}{area in square metres}
}
\description{
Convert Yearly Height (mm) to Volume Flow (unit?)
}
