% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyHeightToVolumeFlow.R
\name{yearlyHeightToVolumeFlow}
\alias{yearlyHeightToVolumeFlow}
\title{Convert Yearly Height (mm) to Volume Flow (unit?)}
\usage{
yearlyHeightToVolumeFlow(height, area)
}
\arguments{
\item{height}{height in mm}

\item{area}{area in square metres}
}
\description{
Convert Yearly Height (mm) to Volume Flow (unit?)
}
