% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realEvapoTranspiration.R
\name{yRatio_3}
\alias{yRatio_3}
\title{Lookup yRatio for given xRatio on a BAGROV curve}
\usage{
yRatio_3(
  bagrovParameter,
  xRatio,
  minSizeForParallel = 10L,
  useAbimoAlgorithm = FALSE
)
}
\arguments{
\item{bagrovParameter}{(vector of) BAGROV parameter(s)}

\item{xRatio}{(vector of) x-ratio(s) (between precipitation and potential
evaporation) for which to look up the corresponding y-ratio(s) (between
actual evaporation and potential evaporation) on the BAGROV curve(s)}

\item{minSizeForParallel}{minimum number of BAGROV curves to start
parallel processing}

\item{useAbimoAlgorithm}{whether or not to use the original algorithm that
is implemented in the C++ code (converted to R:
\code{kwb.rabimo:::yratio_cpp}). Default: \code{FALSE}}
}
\description{
Lookup yRatio for given xRatio on a BAGROV curve
}
