% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_config_to_config.R
\name{abimo_config_to_config}
\alias{abimo_config_to_config}
\title{Convert Abimo Configuration to List}
\usage{
abimo_config_to_config(abimo_config)
}
\arguments{
\item{abimo_config}{as returned by \code{kwb.abimo:::read_config}}
}
\value{
list with elements \code{"potentialEvaporation"},
\code{"runoffFactors"}, \code{"bagrovValues"}, \code{"diverse"},
\code{"resultDigits"}
}
\description{
Convert Abimo Configuration to List
}
