% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualEvaporation.R
\name{actualEvaporationWaterbodyOrPervious}
\alias{actualEvaporationWaterbodyOrPervious}
\title{Calculate Actual Evapotranspiration for Waterbodies or Pervious Areas}
\usage{
actualEvaporationWaterbodyOrPervious(
  usageTuple,
  potentialEvaporation,
  soilProperties,
  precipitation,
  dbg = TRUE,
  ...
)
}
\arguments{
\item{usageTuple}{list as returned by \code{\link{getUsageTuple}}}

\item{potentialEvaporation}{potential evaporation in mm}

\item{soilProperties}{list as returned by \code{\link{getSoilProperties}}}

\item{precipitation}{precipitation in mm}

\item{dbg}{logical indicating whether or not to show debug messages}

\item{\dots}{further arguments passed to \code{\link{realEvapoTranspiration}}
such as \code{runParallel}, \code{blocksize}}
}
\description{
Calculate Actual Evapotranspiration for Waterbodies or Pervious Areas
}
