% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_delta_w.R
\name{calculate_delta_w}
\alias{calculate_delta_w}
\title{Deviation from Natural Water Balance (Delta-W)}
\usage{
calculate_delta_w(
  natural,
  urban,
  columns_water_balance = c("runoff", "infiltr", "evapor"),
  column_code = "code",
  digits = 1L
)
}
\arguments{
\item{natural}{R-Abimo results for the "natural" scenario}

\item{urban}{R-Abimo results for the "urban" scenario}

\item{columns_water_balance}{names of columns in \code{natural} and
\code{urban}, respectively, containing the water balance components runoff,
infiltration, evaporation. Default: \code{c("runoff", "infiltr", "evapor")}}

\item{column_code}{name of column in \code{natural} and \code{urban},
respectively, containing the block area identifiers.}

\item{digits}{integer indicating the number of decimal places in the result}
}
\value{
a data frame with the area codes in column \code{code} and the
delta-W values in column \code{delta_w}
}
\description{
Calculate the deviation from the natural water balance (delta-W) given
R-Abimo results as returned by \code{\link{run_rabimo}}.
}
