% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-scenario.R
\name{calculate_delta_W}
\alias{calculate_delta_W}
\title{Deviation from Natural Water Balance (Delta-W)}
\usage{
calculate_delta_W(natural, urban, ..., implementation = 3L)
}
\arguments{
\item{natural}{R-Abimo results for the "natural" scenario}

\item{urban}{R-Abimo results for the "urban" scenario}

\item{\dots}{further arguments passed to the implementation function as
specified in \code{implementation}}

\item{implementation}{one of 1, 2, 3, indicating a different implementation}
}
\value{
a data frame with the area codes in column \code{code} and the
delta-W values in column \code{delta_w}
}
\description{
Calculate the deviation from the natural water balance (delta-W) given
R-Abimo results as returned by \code{\link{run_rabimo}}.
}
