% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_with_data}
\alias{call_with_data}
\title{Call a Function with Argument Combinations from a Data Frame}
\usage{
call_with_data(FUN, data, ..., simplify = TRUE)
}
\arguments{
\item{FUN}{function to be called}

\item{data}{data frame with one column per argument of \code{FUN}}

\item{\dots}{further arguments passed to \code{\link{mapply}} via
\code{MoreArgs}}

\item{simplify}{passed to \code{\link{mapply}}}
}
\value{
vector of length \code{nrow(data)} with the result values returned by
\code{FUN}
}
\description{
Call a Function with Argument Combinations from a Data Frame
}
\examples{
combis <- expand.grid(x = 1:2, y = c(10, 20, 30))
combis

call_with_data(`+`, combis)
}
