% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{crop_box}
\alias{crop_box}
\title{Crop a box out of a shape}
\usage{
crop_box(x, xoffset = 0.45, yoffset = 0.45, xscale = 0.1, yscale = 0.1)
}
\arguments{
\item{x}{sf object}

\item{xoffset}{x-offset as fraction of original width (0..1)}

\item{yoffset}{y-offset as fraction of original height (0..1)}

\item{xscale}{new width as fraction of original width (0..1)}

\item{yscale}{new height as fraction of original height (0..1)}
}
\description{
Crop a box out of a shape
}
