% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-scenario.R
\name{data_to_natural}
\alias{data_to_natural}
\title{Transform R-Abimo input Data into their natural scenario equivalent}
\usage{
data_to_natural(data, type = "undeveloped")
}
\arguments{
\item{data}{The data Set in R-Abimo format}

\item{type}{a character object containing the name of natural scenario. Defaults to "undeveloped"}
}
\value{
a dataframe with R-Abimo input data for the chosen natural scenario
}
\description{
Three scenarios are possible:
\enumerate{
\item undeveloped: all paved or constructed areas are set to 0\%. No connection to the sewer.
\item forested: like undeveloped, but the land type is declared to be "forested".
\item horticultural: like undeveloped, but the land type is declared to be "horticultural".
}
}
