% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{define_controls}
\alias{define_controls}
\title{Define List of "Controls"}
\usage{
define_controls(
  check = TRUE,
  use_abimo_bagrov_solver = TRUE,
  reproduce_abimo_error = FALSE,
  output_format = "rabimo",
  intermediates = FALSE
)
}
\arguments{
\item{check}{logical indicating whether the check functions are executed.
Default: \code{TRUE}.}

\item{use_abimo_bagrov_solver}{logical indicating whether or not to use the
(fast!) algorithm implemented in Abimo to solve the Bagrov equations.
Default: \code{TRUE}.}

\item{reproduce_abimo_error}{logical indicating whether or not to reproduce
the error that is contained in Abimo (missing area fraction factor).
Default: \code{FALSE}.}

\item{output_format}{one of "abimo" (upper case columns: CODE, R, ROW, RI,
RVOL, ROWVOL, RIVOL, FLAECHE, VERDUNSTUN), "rabimo" (lower case columns:
code, surface_runoff, infiltration, evaporation).}

\item{intermediates}{logical indicating whether the intermediate results are
returned as attributes. Default: \code{FALSE}.}
}
\value{
list with the arguments of this function as list elements
}
\description{
Define a list of settings that control how the main function
\code{\link{run_rabimo}} should behave.
}
\examples{
inputs <- kwb.rabimo::rabimo_inputs_2020
test_data <- inputs$data[sample(seq_len(nrow(inputs$data)), size = 1000L), ]
controls_default <- define_controls()
controls_no_check <- define_controls(check = FALSE)
controls_no_solver <- define_controls(use_abimo_bagrov_solver = FALSE)
system.time(result_default <- kwb.rabimo::run_rabimo(
  test_data, inputs$config, controls_default
))
system.time(result_no_check <- kwb.rabimo::run_rabimo(
  test_data, inputs$config, controls_no_check
))
identical(result_default, result_no_check)
\dontrun{
system.time(result_no_solver <- kwb.rabimo::run_rabimo(
  test_data, inputs$config, controls_no_solver
))
}
}
