% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_measures.R
\name{distribute_measures}
\alias{distribute_measures}
\title{Distribute Rainwater Management Measures to Single Block Areas}
\usage{
distribute_measures(blocks, targets, intermediates = FALSE)
}
\arguments{
\item{blocks}{data frame as being input to \code{\link{run_rabimo}}}

\item{targets}{list with element \code{green_roof}, \code{unpaved},
\code{to_swale}, each of which is a value between 0 and 1, describing the
extent to which the corresponding measures are to be installed on average
over all \code{blocks}}

\item{intermediates}{logical indicating whether or not to return tables with
intermediate results in attributes. The default is \code{FALSE}, i.e. no
attributes are attached to the result data frame.}
}
\value{
data frame with the columns describing the measurements being
updated. In case of \code{intermediates = TRUE}, the data frame has
attributes \code{green_roof_table}, \code{unpaved_area_table},
\code{swale_connection_table}, carrying intermediate results.
}
\description{
Distribute Rainwater Management Measures to Single Block Areas
}
