% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rabimo_area.R
\name{generate_rabimo_area}
\alias{generate_rabimo_area}
\title{Generate an area in R-Abimo format with default values}
\usage{
generate_rabimo_area(code = NULL, ..., column_info = read_column_info())
}
\arguments{
\item{code}{vector of unique area identifiers. If NULL, default codes are
created: area_1, area_2, ...}

\item{\dots}{key = value pairs overriding the default column values}

\item{column_info}{data frame as returned by \code{\link{read_column_info}}}

\item{n}{number of areas to be created. Default: length of code vector (1 if code is NULL)}
}
\description{
All default values can be overridden by entering new key-value pairs.
}
