% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rabimo_area.R
\name{generate_rabimo_area}
\alias{generate_rabimo_area}
\title{Generate an area in R-ABIMO format with default values.
All default values can be overridden by entering new key-value pairs.}
\usage{
generate_rabimo_area(code, ..., column_info = read_column_info())
}
\arguments{
\item{\dots}{key = value pairs overriding the default column values}

\item{column_info}{data frame as returned by \code{\link{read_column_info}}}
}
\description{
Generate an area in R-ABIMO format with default values.
All default values can be overridden by entering new key-value pairs.
}
