% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPotentialEvaporation.R
\name{getPotentialEvaporation}
\alias{getPotentialEvaporation}
\title{Provide Data on Potential Evaporation}
\usage{
getPotentialEvaporation(isWaterbody, district, lookup)
}
\arguments{
\item{isWaterbody}{(vector of) logical indicating whether a block area is
of type (from the type/yield/irrigation tuple) "waterbody"}

\item{district}{(vector of) integer indicating the district number of the
plot area (from the original input column "BEZIRK")}

\item{lookup}{data frame with key columns \code{isWaterbody}, \code{district}
and value columns \code{etp}, \code{etps}. A data frame of the required
structure is returned by \code{\link{abimo_config_to_config}} in list
element \code{"potentialEvaporation"}}
}
\description{
Provide Data on Potential Evaporation
}
\examples{
\dontrun{
config <- abimo_config_to_config(kwb.abimo:::read_config())
getPotentialEvaporation(
  is_waterbody = TRUE,
  district = 1,
  lookup = config$potentialEvaporation
)
}

}
