% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPotentialEvaporation.R
\name{getPotentialEvaporation}
\alias{getPotentialEvaporation}
\title{Provide Data on Potential Evaporation}
\usage{
getPotentialEvaporation(isWaterbody, district, config)
}
\arguments{
\item{isWaterbody}{(vector of) logical indicating whether a block area is
of type (from the type/yield/irrigation tuple) "waterbody"}

\item{district}{(vector of) integer indicating the district number of the
plot area (from the original input column "BEZIRK")}

\item{config}{list structure as returned by
\code{\link{abimo_config_to_config}}}
}
\description{
Provide Data on Potential Evaporation
}
\examples{
\dontrun{
config <- abimo_config_to_config(kwb.abimo:::read_config())
getPotentialEvaporation(
  is_waterbody = TRUE,
  district = 1,
  config = config
)
}

}
