% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_block.R
\name{plot_block}
\alias{plot_block}
\title{Plot Area Fractions and their Names for one Block}
\usage{
plot_block(block, cex = 1, delta = 0.1)
}
\arguments{
\item{block}{data frame with one row and columns as provided in the data
frame \code{\link{rabimo_inputs_2025}$data}}

\item{cex}{character expansion factor to scale the texts}

\item{delta}{controls the space between rectangles}
}
\description{
Plot Area Fractions and their Names for one Block
}
