% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabimo_inputs_2025.R
\docType{data}
\name{rabimo_inputs_2025}
\alias{rabimo_inputs_2025}
\title{R-Abimo Inputs (Data and Config) for Berlin, 2025.}
\format{
\subsection{\code{rabimo_inputs_2025}}{

A list containing two elements:
\describe{
\item{data}{a data frame with the input data in r-abimo format ...(number of vars)}
\item{config}{a list object with configuration data}
}
}

\subsection{\code{rabimo_inputs_2025$data}}{

A data.frame with 58531 observations of 25 variables:
\describe{
\item{\code{code}}{Unique block area identifier (character)}
\item{\code{prec_yr}}{Long-term average of annual precipitation in mm (integer)}
\item{\code{prec_s}}{Long-term average of annual precipitation within summer months (May to October) in mm (integer)}
\item{\code{epot_yr}}{Long-term average of annual potential evapotranspiration in mm (integer)}
\item{\code{epot_s}}{Long-term average of annual potential evapotranspiration within summer months (May to October) in mm (integer)}
\item{\code{district}}{Number of Berlin "Bezirk" (district) in which the block area is located (character)}
\item{\code{total_area}}{Total block area in square metres (numeric)}
\item{\code{roof}}{Fraction of the total area that is considered as "roof" area (numeric value between 0.0 and 1.0)}
\item{\code{green_roof}}{Fraction of the roof area that belongs to green roofs (numeric value between 0.0 and 1.0). A value of 1.0 means that all roofs in the block area are green roofs.}
\item{\code{swg_roof}}{Fraction of the roof area that is connected to the drainage system (numeric value between 0.0 and 1.0)}
\item{\code{pvd}}{Fraction of the total block area that is paved (numeric value between 0.0 and 1.0)}
\item{\code{swg_pvd}}{Fraction of the paved area that is connected to the drainage system (numeric value between 0.0 and 1.0)}
\item{\code{srf1_pvd}}{Fraction of the paved area that belongs to surface class 1 (numeric value between 0.0 and 1.0, see note 1 below)}
\item{\code{srf2_pvd}}{Fraction of the paved area that belongs to surface class 2 (numeric value between 0.0 and 1.0, see note 1 below)}
\item{\code{srf3_pvd}}{Fraction of the paved area that belongs to surface class 3 (numeric value between 0.0 and 1.0, see note 1 below)}
\item{\code{srf4_pvd}}{Fraction of the paved area that belongs to surface class 4 (numeric value between 0.0 and 1.0, see note 1 below)}
\item{\code{srf5_pvd}}{Fraction of the paved area that belongs to surface class 5 (numeric value between 0.0 and 1.0, see note 1 below)}
\item{\code{to_swale}}{Fraction of sealed area (roof area + paved area) that is connected to an infiltration swale (numeric)}
\item{\code{gw_dist}}{Distance between groundwater table and surface in metres (numeric)}
\item{\code{ufc30}}{field capacity in 30 cm depth (numeric)}
\item{\code{ufc150}}{field capacity in 150 cm depth (numeric)}
\item{\code{land_type}}{land type, one of \code{forested}, \code{horticultural}, \code{urban}, \code{vegetationless}, \code{waterbody} (character)}
\item{\code{veg_class}}{vegetation class index (numeric), derived from an analysis tree volumes}
\item{\code{irrigation}}{irrigation in mm per year (integer)}
\item{\code{block_type}}{Block type identifier of the form "usage-type-id_block-type-id_usage-type-description_block-type-description" (character)}
}
}

\subsection{\code{rabimo_inputs_2025$config}}{

A list with 3 named elements:
\describe{
\item{runoff_factors}{Runoff factors, vector of numeric with names \code{roof}, \code{surface1}, \code{surface2}, \code{surface3}, \code{surface4}, \code{surface5}}
\item{bagrov_values}{Bagrov values for sealed surfaces, vector of numeric with names \code{roof}, \code{green_roof}, \code{surface1}, \code{surface2}, \code{surface3}, \code{surface4}, \code{surface5}}
\item{swale}{Model parameter(s) related to the 'swale' measure, vector of numeric with currently one value, named \code{swale_evaporation_factor}}
}
}
}
\source{
\url{https://www.berlin.de/umweltatlas/en/general/contact/}

\url{https://gdi.berlin.de/services/wfs/ua_gruendaecher_2020?REQUEST=GetCapabilities&SERVICE=wfs}
}
\usage{
rabimo_inputs_2025
}
\description{
Data frame and configuration object that are required by the R-Abimo main
function \code{\link{run_rabimo}}.
The data have been provided by
Senate Department for Urban Development, Building and Housing
III D Spatial Data Infrastructure, Environmental Atlas.
}
\keyword{datasets}
