% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_column_info.R
\name{read_column_info}
\alias{read_column_info}
\title{Provide Meta Information About Input Columns}
\usage{
read_column_info()
}
\value{
data frame with columns "rabimo_berlin", "abimo_berlin", "by_100",
"meaning", "unit", "type", "data_type", "default"
}
\description{
Provide Meta Information About Input Columns
}
