% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realEvapoTranspiration.R
\name{realEvapoTranspiration}
\alias{realEvapoTranspiration}
\title{Calculate Actual Evapotranspiration with Bagrov}
\usage{
realEvapoTranspiration(
  precipitation,
  potentialEvaporation,
  bagrovParameter,
  xRatio = NULL,
  FUN_yRatio = yRatio_3,
  ...
)
}
\arguments{
\item{precipitation}{precipitation in mm}

\item{potentialEvaporation}{potential evaporation in mm}

\item{bagrovParameter}{Bagrov parameter (n-value)}

\item{xRatio}{optional. Instead of \code{precipitation} and
\code{potentialEvaporation} the quotient of both may be passed to this
function. The idea is to calculate the quotient out of the function and to
reuse the quotient instead of recalculating it.}

\item{FUN_yRatio}{function to be called to calculate the yRatio(s) from the
given xRatio(s). Default: \code{kwb.rabimo:::yRatio_3}}

\item{\dots}{further arguments passed to \code{FUN_yRatio}}
}
\value{
estimated actual evapotranspiration in mm
}
\description{
Calculate Actual Evapotranspiration with Bagrov
}
