% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{run_rabimo}
\alias{run_rabimo}
\title{Run R-Abimo, the R-implementation of Water Balance Model Abimo}
\usage{
run_rabimo(data, config, controls = define_controls())
}
\arguments{
\item{data}{data frame with columns as returned by
\code{\link{prepare_input_data}}}

\item{config}{configuration object (list) as returned by
\code{\link{abimo_config_to_config}}}

\item{controls}{list of settings that control how the function should behave.
Use \code{\link{define_controls}} to define such a list. The default is
the list returned by \code{define_controls()}.}
}
\value{
data frame with columns as returned by Abimo
}
\description{
Run R-Abimo, the R-implementation of Water Balance Model Abimo
}
