% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{run_rabimo}
\alias{run_rabimo}
\title{Run R-Abimo, the R-implementation of Water Balance Model Abimo}
\usage{
run_rabimo(input_data, config)
}
\arguments{
\item{input_data}{data frame with columns as required by Abimo}

\item{config}{configuration object (list) as returned by
\code{kwb.abimo:::read_config()}}
}
\value{
data frame with columns as returned by Abimo
}
\description{
Run R-Abimo, the R-implementation of Water Balance Model Abimo
}
