% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo.R
\name{run_rabimo}
\alias{run_rabimo}
\title{Run R-Abimo, the R-implementation of Water Balance Model Abimo}
\usage{
run_rabimo(data, config, simulate_abimo = TRUE)
}
\arguments{
\item{data}{data frame with columns as returned by
\code{\link{prepare_input_data}}}

\item{config}{configuration object (list) as returned by
\code{abimo_config_to_config()}, used on \code{kwb.abimo::read_config()}}

\item{simulate_abimo}{logical indicating whether or not to simulate exactly
what Abimo does (including obvious errors!).
Default: \code{TRUE}!}
}
\value{
data frame with columns as returned by Abimo
}
\description{
Run R-Abimo, the R-implementation of Water Balance Model Abimo
}
