% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo_with_measures.R
\name{run_rabimo_with_measures}
\alias{run_rabimo_with_measures}
\title{Distribute Rainwater Management Measures and run R-Abimo}
\usage{
run_rabimo_with_measures(
  blocks,
  measures,
  config = kwb.rabimo::rabimo_inputs_2020$config,
  old_version = FALSE,
  ...
)
}
\arguments{
\item{blocks}{data frame of selected blocks (same columns as in
\code{\link{rabimo_inputs_2020}$data})}

\item{measures}{list with elements \code{green_roof}, \code{unpaved},
\code{to_swale} representing the target percentages of the total areas
corresponding to each measure}

\item{config}{configuration object, default:
\code{\link{rabimo_inputs_2020}$config}}

\item{old_version}{if \code{TRUE} the old, erroneous version of this function
is used (not correctly considering the updated pvd value before calculating
the new to_swale values). The default is \code{FALSE}.}

\item{\dots}{further arguments passed to \code{\link{run_rabimo}}, such as
\code{silent = TRUE}}
}
\description{
Distribute Rainwater Management Measures and run R-Abimo
}
