% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rabimo_with_measures.R
\name{run_rabimo_with_measures}
\alias{run_rabimo_with_measures}
\title{Distribute Rainwater Management Measures and run R-Abimo}
\usage{
run_rabimo_with_measures(
  blocks,
  measures,
  config = kwb.rabimo::rabimo_inputs_2020$config
)
}
\arguments{
\item{blocks}{data frame of selected blocks (same columns as in
\code{link{rabimo_inputs_2020}$data})}

\item{measures}{list with elements \code{green_roof}, \code{unpaved},
\code{to_swale} representing the target percentages of the total areas
corresponding to each measure}
}
\description{
Distribute Rainwater Management Measures and run R-Abimo
}
