% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_of_fractions.R
\name{triangle_of_fractions}
\alias{triangle_of_fractions}
\title{Plot Triangle of Three Fractions}
\usage{
triangle_of_fractions(fractions, cols = c("blue", "red", "green"))
}
\arguments{
\item{fractions}{numeric vector with three values having a sum of one. The
names of the vector elements are used as labels}

\item{cols}{vector of length three giving the colour names}
}
\description{
Plot Triangle of Three Fractions
}
\examples{
# blue, red, green | left, right, bottom
components <- c(runoff = 200, infiltration = 50, evaporation = 100)
fractions <- components / sum(components)
triangle_of_fractions(fractions)
}
