% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realEvapoTranspiration.R
\name{yRatio_3}
\alias{yRatio_3}
\title{Lookup yRatio for given xRatio on a BAGROV curve}
\usage{
yRatio_3(
  bagrovParameter,
  xRatio,
  runParallel = length(bagrovParameter) > 10L,
  blocksize = 10L
)
}
\arguments{
\item{runParallel}{logical indicating whether or not parallel processing is
to be used. By default it is used if more than 1000 values are to be
calculated}

\item{blocksize}{in case of parallel processing this is the size of the
data frames that are processed by \code{\link{call_with_data}} on different
cores}
}
\description{
Lookup yRatio for given xRatio on a BAGROV curve
}
